@@ -11,13 +11,13 @@
 // Software description: A SwiftUI components library with code examples for Orange Unified Design System
 //
 
-import Foundation
 import OUDSTokensRaw
+import SwiftUICore
 
 /// Kind of semantic tokens which will wrap a combination of `TypographyCompositeRawToken` depending to size classes.
 /// Kind of composite token with multiple values, but not named "composite" because this word is already used in the design system.
 /// Allows to gather the multiple-value tokens from Figma inside one object.
-public final class MultipleTypographyTokens: NSObject {
+public final class MultipleTypographyTokens: NSObject, Sendable {
 
     /// For **extra-compact** and **compact** viewports
     public let compact: TypographyCompositeRawToken
@@ -49,5 +49,12 @@ public final class MultipleTypographyTokens: NSObject {
     override public func isEqual(_ object: Any?) -> Bool {
         guard let object = object as? MultipleTypographyTokens else { return false }
         return self.compact == object.compact && self.regular == object.regular
+    }
+
+    /// Returns the composite raw token of typography to use according to the `userInterfaceSizeClass` (i.e. `compact` or `regular`)
+    /// - Parameter userInterfaceSizeClass: The user interface size class
+    /// - Returns: The composite raw token to use (of type `TypographyCompositeRawToken`)
+    public func typographyToken(for userInterfaceSizeClass: UserInterfaceSizeClass) -> TypographyCompositeRawToken {
+        userInterfaceSizeClass == .compact ? compact : regular
     }
 }
