@@ -15,17 +15,11 @@ import Foundation
 import OUDSTokensRaw
 import OUDSTokensSemantic
 
-// ଘ( ･ω･)_/ﾟ･:*:･｡☆
-// [File to generate with the tokenator]
-
-/// Defines basic values common to all themes for `OpacitySemanticTokens`.
-/// These values can be overriden inside `OUDSTheme` subclasses (in extensions or not, in the same module or not) thanks to the `@objc open` combination.
 extension OUDSTheme: OpacitySemanticTokens {
-
-    @objc open var opacityTransparent: OpacitySemanticToken { OpacityRawTokens.opacity0 }
-    @objc open var opacityWeaker: OpacitySemanticToken { OpacityRawTokens.opacity100 }
-    @objc open var opacityWeak: OpacitySemanticToken { OpacityRawTokens.opacity300 }
     @objc open var opacityMedium: OpacitySemanticToken { OpacityRawTokens.opacity500 }
-    @objc open var opacityStrong: OpacitySemanticToken { OpacityRawTokens.opacity700 }
     @objc open var opacityOpaque: OpacitySemanticToken { OpacityRawTokens.opacity900 }
+    @objc open var opacityStrong: OpacitySemanticToken { OpacityRawTokens.opacity700 }
+    @objc open var opacityTransparent: OpacitySemanticToken { OpacityRawTokens.opacity0 }
+    @objc open var opacityWeak: OpacitySemanticToken { OpacityRawTokens.opacity300 }
+    @objc open var opacityWeaker: OpacitySemanticToken { OpacityRawTokens.opacity100 }
 }
