@@ -67,6 +67,21 @@ final class TestThemeOverrideOfTypographySemanticTokens: XCTestCase {
 
     // MARK: - Semantic token - Typography - Font - Weight
 
+    func testInheritedThemeCanOverrideSemanticTokenFontWeightDefault() throws {
+        XCTAssertNotEqual(inheritedTheme.fontWeightDefault, abstractTheme.fontWeightDefault)
+        XCTAssertTrue(inheritedTheme.fontWeightDefault == MockTheme.mockThemeTypographyFontWeightRawToken)
+    }
+
+    func testInheritedThemeCanOverrideSemanticTokenFontWeightCode() throws {
+        XCTAssertNotEqual(inheritedTheme.fontWeightCode, abstractTheme.fontWeightCode)
+        XCTAssertTrue(inheritedTheme.fontWeightCode == MockTheme.mockThemeTypographyFontWeightRawToken)
+    }
+
+    func testInheritedThemeCanOverrideSemanticTokenFontWeightStrong() throws {
+        XCTAssertNotEqual(inheritedTheme.fontWeightStrong, abstractTheme.fontWeightStrong)
+        XCTAssertTrue(inheritedTheme.fontWeightStrong == MockTheme.mockThemeTypographyFontWeightRawToken)
+    }
+
     func testInheritedThemeCanOverrideSemanticTokenFontWeightDisplay() throws {
         XCTAssertNotEqual(inheritedTheme.fontWeightDisplay, abstractTheme.fontWeightDisplay)
         XCTAssertTrue(inheritedTheme.fontWeightDisplay == MockTheme.mockThemeTypographyFontWeightRawToken)
@@ -95,11 +110,6 @@ final class TestThemeOverrideOfTypographySemanticTokens: XCTestCase {
     func testInheritedThemeCanOverrideSemanticTokenFontWeightLabelStrong() throws {
         XCTAssertNotEqual(inheritedTheme.fontWeightLabelStrong, abstractTheme.fontWeightLabelStrong)
         XCTAssertTrue(inheritedTheme.fontWeightLabelStrong == MockTheme.mockThemeTypographyFontWeightRawToken)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenFontWeightCode() throws {
-        XCTAssertNotEqual(inheritedTheme.fontWeightCode, abstractTheme.fontWeightCode)
-        XCTAssertTrue(inheritedTheme.fontWeightCode == MockTheme.mockThemeTypographyFontWeightRawToken)
     }
 
     // MARK: - Semantic token - Typography - Font - Size
@@ -348,131 +358,6 @@ final class TestThemeOverrideOfTypographySemanticTokens: XCTestCase {
     func testInheritedThemeCanOverrideSemanticTokenFontLineHeightCodeSmall() throws {
         XCTAssertNotEqual(inheritedTheme.fontLineHeightCodeSmall, abstractTheme.fontLineHeightCodeSmall)
         XCTAssertTrue(inheritedTheme.fontLineHeightCodeSmall == MockTheme.mockThemeMultipleTypographyLineHeightTokens)
-    }
-
-    // MARK: - Semantic tokens - Typography - Composites - Display
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeDisplayLarge() throws {
-        XCTAssertNotEqual(inheritedTheme.typeDisplayLarge, abstractTheme.typeDisplayLarge)
-        XCTAssertTrue(inheritedTheme.typeDisplayLarge == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeDisplayMedium() throws {
-        XCTAssertNotEqual(inheritedTheme.typeDisplayMedium, abstractTheme.typeDisplayMedium)
-        XCTAssertTrue(inheritedTheme.typeDisplayMedium == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeDisplaySmal() throws {
-        XCTAssertNotEqual(inheritedTheme.typeDisplaySmall, abstractTheme.typeDisplaySmall)
-        XCTAssertTrue(inheritedTheme.typeDisplaySmall == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    // MARK: - Semantic tokens - Typography - Composites - Heading
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeHeadingXLarge() throws {
-        XCTAssertNotEqual(inheritedTheme.typeHeadingXLarge, abstractTheme.typeHeadingXLarge)
-        XCTAssertTrue(inheritedTheme.typeHeadingXLarge == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeHeadingLarge() throws {
-        XCTAssertNotEqual(inheritedTheme.typeHeadingLarge, abstractTheme.typeHeadingLarge)
-        XCTAssertTrue(inheritedTheme.typeHeadingLarge == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeHeadingMedium() throws {
-        XCTAssertNotEqual(inheritedTheme.typeHeadingMedium, abstractTheme.typeHeadingMedium)
-        XCTAssertTrue(inheritedTheme.typeHeadingMedium == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeHeadingSmall() throws {
-        XCTAssertNotEqual(inheritedTheme.typeHeadingSmall, abstractTheme.typeHeadingSmall)
-        XCTAssertTrue(inheritedTheme.typeHeadingSmall == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    // MARK: - Semantic tokens - Typography - Composites - Body
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeDefaultLarge() throws {
-        XCTAssertNotEqual(inheritedTheme.typeBodyDefaultLarge, abstractTheme.typeBodyDefaultLarge)
-        XCTAssertTrue(inheritedTheme.typeBodyDefaultLarge == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeDefaultMedium() throws {
-        XCTAssertNotEqual(inheritedTheme.typeBodyDefaultMedium, abstractTheme.typeBodyDefaultMedium)
-        XCTAssertTrue(inheritedTheme.typeBodyDefaultMedium == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeDefaultSmall() throws {
-        XCTAssertNotEqual(inheritedTheme.typeBodyDefaultSmall, abstractTheme.typeBodyDefaultSmall)
-        XCTAssertTrue(inheritedTheme.typeBodyDefaultSmall == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeStrongLarge() throws {
-        XCTAssertNotEqual(inheritedTheme.typeBodyStrongLarge, abstractTheme.typeBodyStrongLarge)
-        XCTAssertTrue(inheritedTheme.typeBodyStrongLarge == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeStrongMedium() throws {
-        XCTAssertNotEqual(inheritedTheme.typeBodyStrongMedium, abstractTheme.typeBodyStrongMedium)
-        XCTAssertTrue(inheritedTheme.typeBodyStrongMedium == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeStrongSmall() throws {
-        XCTAssertNotEqual(inheritedTheme.typeBodyStrongSmall, abstractTheme.typeBodyStrongSmall)
-        XCTAssertTrue(inheritedTheme.typeBodyStrongSmall == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    // MARK: - Semantic tokens - Typography - Composites - Label
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeLabelDefaultXLarge() throws {
-        XCTAssertNotEqual(inheritedTheme.typeLabelDefaultXLarge, abstractTheme.typeLabelDefaultXLarge)
-        XCTAssertTrue(inheritedTheme.typeLabelDefaultXLarge == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeLabelDefaultLarge() throws {
-        XCTAssertNotEqual(inheritedTheme.typeLabelDefaultLarge, abstractTheme.typeLabelDefaultLarge)
-        XCTAssertTrue(inheritedTheme.typeLabelDefaultLarge == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeLabelDefaultMedium() throws {
-        XCTAssertNotEqual(inheritedTheme.typeLabelDefaultMedium, abstractTheme.typeLabelDefaultMedium)
-        XCTAssertTrue(inheritedTheme.typeLabelDefaultMedium == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeLabelDefaultSmall() throws {
-        XCTAssertNotEqual(inheritedTheme.typeLabelDefaultSmall, abstractTheme.typeLabelDefaultSmall)
-        XCTAssertTrue(inheritedTheme.typeLabelDefaultSmall == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeLabelStrongXLarge() throws {
-        XCTAssertNotEqual(inheritedTheme.typeLabelStrongXLarge, abstractTheme.typeLabelStrongXLarge)
-        XCTAssertTrue(inheritedTheme.typeLabelStrongXLarge == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeLabelStrongLarge() throws {
-        XCTAssertNotEqual(inheritedTheme.typeLabelStrongLarge, abstractTheme.typeLabelStrongLarge)
-        XCTAssertTrue(inheritedTheme.typeLabelStrongLarge == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeLabelStrongMedium() throws {
-        XCTAssertNotEqual(inheritedTheme.typeLabelStrongMedium, abstractTheme.typeLabelStrongMedium)
-        XCTAssertTrue(inheritedTheme.typeLabelStrongMedium == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeLabelStrongSmall() throws {
-        XCTAssertNotEqual(inheritedTheme.typeLabelStrongSmall, abstractTheme.typeLabelStrongSmall)
-        XCTAssertTrue(inheritedTheme.typeLabelStrongSmall == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    // MARK: - Semantic tokens - Typography - Composites - Code
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeCodeMedium() throws {
-        XCTAssertNotEqual(inheritedTheme.typeCodeMedium, abstractTheme.typeCodeMedium)
-        XCTAssertTrue(inheritedTheme.typeCodeMedium == MockTheme.mockThemeMultipleTypographyTokens)
-    }
-
-    func testInheritedThemeCanOverrideSemanticTokenTypeCodeSmall() throws {
-        XCTAssertNotEqual(inheritedTheme.typeCodeSmall, abstractTheme.typeCodeSmall)
-        XCTAssertTrue(inheritedTheme.typeCodeSmall == MockTheme.mockThemeMultipleTypographyTokens)
     }
 }
 
