@@ -2,45 +2,73 @@
 // Software Name: OUDS iOS
 // SPDX-FileCopyrightText: Copyright (c) Orange SA
 // SPDX-License-Identifier: MIT
-// 
+//
 // This software is distributed under the MIT license,
 // the text of which is available at https://opensource.org/license/MIT/
 // or see the "LICENSE" file for more details.
-// 
-// Authors: See CONTRIBUTORS.txt
-// Software description: A SwiftUI components library with code examples for Orange Unified Design System 
 //
-
-// ଘ( ･ω･)_/ﾟ･:*:･｡☆
-// [File to generate with the tokenator]
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
 
 // swiftlint:disable missing_docs
 
-/// Extracted in a separated file to help the *Figma* JSON to Swift parser to generate files to include easily.
 extension ColorRawTokens {
-
-    // MARK: Primitive token - Colors - Black, white
-
-    public static let colorFunctionalWhite: ColorRawToken = "#FFFFFF"
+    public static let colorDecorativeAmber100: ColorRawToken = "#FFF0CC"
+    public static let colorDecorativeAmber200: ColorRawToken = "#FFE199"
+    public static let colorDecorativeAmber300: ColorRawToken = "#FFD266"
+    public static let colorDecorativeAmber400: ColorRawToken = "#FFC333"
+    public static let colorDecorativeAmber500: ColorRawToken = "#FFB400"
+    public static let colorDecorativeAmber600: ColorRawToken = "#CC9000"
+    public static let colorDecorativeAmber700: ColorRawToken = "#996C00"
+    public static let colorDecorativeAmber800: ColorRawToken = "#664800"
+    public static let colorDecorativeAmber900: ColorRawToken = "#332400"
+    public static let colorDecorativeAmethyst100: ColorRawToken = "#F1ECF9"
+    public static let colorDecorativeAmethyst200: ColorRawToken = "#E0D4F2"
+    public static let colorDecorativeAmethyst300: ColorRawToken = "#C5ADE6"
+    public static let colorDecorativeAmethyst400: ColorRawToken = "#A885D8"
+    public static let colorDecorativeAmethyst500: ColorRawToken = "#8D60CD"
+    public static let colorDecorativeAmethyst600: ColorRawToken = "#5B2F98"
+    public static let colorDecorativeAmethyst700: ColorRawToken = "#432371"
+    public static let colorDecorativeAmethyst800: ColorRawToken = "#2C174A"
+    public static let colorDecorativeAmethyst900: ColorRawToken = "#150B23"
+    public static let colorDecorativeDeepPeach100: ColorRawToken = "#FBEBDF"
+    public static let colorDecorativeDeepPeach200: ColorRawToken = "#F4CFB2"
+    public static let colorDecorativeDeepPeach300: ColorRawToken = "#E3B591"
+    public static let colorDecorativeDeepPeach400: ColorRawToken = "#C19372"
+    public static let colorDecorativeDeepPeach500: ColorRawToken = "#CF7E3F"
+    public static let colorDecorativeDeepPeach600: ColorRawToken = "#AA6631"
+    public static let colorDecorativeDeepPeach700: ColorRawToken = "#7E4F2A"
+    public static let colorDecorativeDeepPeach800: ColorRawToken = "#553720"
+    public static let colorDecorativeDeepPeach900: ColorRawToken = "#2E2014"
+    public static let colorDecorativeEmerald100: ColorRawToken = "#E5F5ED"
+    public static let colorDecorativeEmerald200: ColorRawToken = "#C0E8D4"
+    public static let colorDecorativeEmerald300: ColorRawToken = "#9BDABA"
+    public static let colorDecorativeEmerald400: ColorRawToken = "#75CCA1"
+    public static let colorDecorativeEmerald500: ColorRawToken = "#50BE87"
+    public static let colorDecorativeEmerald600: ColorRawToken = "#3BA06E"
+    public static let colorDecorativeEmerald700: ColorRawToken = "#2E7B54"
+    public static let colorDecorativeEmerald800: ColorRawToken = "#20563B"
+    public static let colorDecorativeEmerald900: ColorRawToken = "#123021"
+    public static let colorDecorativeShockingPink100: ColorRawToken = "#FFE5F6"
+    public static let colorDecorativeShockingPink200: ColorRawToken = "#FFB4E6"
+    public static let colorDecorativeShockingPink300: ColorRawToken = "#FF80D4"
+    public static let colorDecorativeShockingPink400: ColorRawToken = "#FF4DC3"
+    public static let colorDecorativeShockingPink500: ColorRawToken = "#FF1AB2"
+    public static let colorDecorativeShockingPink600: ColorRawToken = "#E50099"
+    public static let colorDecorativeShockingPink700: ColorRawToken = "#B20077"
+    public static let colorDecorativeShockingPink800: ColorRawToken = "#800055"
+    public static let colorDecorativeShockingPink900: ColorRawToken = "#4D0033"
+    public static let colorDecorativeSky100: ColorRawToken = "#D2ECF9"
+    public static let colorDecorativeSky200: ColorRawToken = "#A5DAF3"
+    public static let colorDecorativeSky300: ColorRawToken = "#79C7EC"
+    public static let colorDecorativeSky400: ColorRawToken = "#4AB4E6"
+    public static let colorDecorativeSky500: ColorRawToken = "#1FA2E0"
+    public static let colorDecorativeSky600: ColorRawToken = "#1982B3"
+    public static let colorDecorativeSky700: ColorRawToken = "#136186"
+    public static let colorDecorativeSky800: ColorRawToken = "#0C415A"
+    public static let colorDecorativeSky900: ColorRawToken = "#06202D"
     public static let colorFunctionalBlack: ColorRawToken = "#000000"
-
-    // MARK: Primitive token - Colors - Functional light gray
-
-    public static let colorFunctionalLightGray80: ColorRawToken = "#F4F4F4"
-    public static let colorFunctionalLightGray160: ColorRawToken = "#EEEEEE"
-    public static let colorFunctionalLightGray240: ColorRawToken = "#E0E0E0"
-    public static let colorFunctionalLightGray320: ColorRawToken = "#D6D6D6"
-    public static let colorFunctionalLightGray400: ColorRawToken = "#CCCCCC"
-    public static let colorFunctionalLightGray480: ColorRawToken = "#C2C2C2"
-    public static let colorFunctionalLightGray560: ColorRawToken = "#BBBBBB"
-    public static let colorFunctionalLightGray640: ColorRawToken = "#ADADAD"
-    public static let colorFunctionalLightGray720: ColorRawToken = "#A3A3A3"
-    public static let colorFunctionalLightGray800: ColorRawToken = "#999999"
-    public static let colorFunctionalLightGray880: ColorRawToken = "#8F8F8F"
-    public static let colorFunctionalLightGray960: ColorRawToken = "#858585"
-
-    // MARK: Primitive token - Colors - Functional dark gray
-
     public static let colorFunctionalDarkGray80: ColorRawToken = "#7A7A7A"
     public static let colorFunctionalDarkGray160: ColorRawToken = "#707070"
     public static let colorFunctionalDarkGray240: ColorRawToken = "#666666"
@@ -53,45 +81,6 @@ extension ColorRawTokens {
     public static let colorFunctionalDarkGray800: ColorRawToken = "#1F1F1F"
     public static let colorFunctionalDarkGray880: ColorRawToken = "#141414"
     public static let colorFunctionalDarkGray960: ColorRawToken = "#0A0A0A"
-
-    // MARK: Primitive token - Colors - Functional scarlet
-
-    public static let colorFunctionalScarlet100: ColorRawToken = "#FFE5E6"
-    public static let colorFunctionalScarlet200: ColorRawToken = "#FFB2B3"
-    public static let colorFunctionalScarlet300: ColorRawToken = "#FF8081"
-    public static let colorFunctionalScarlet400: ColorRawToken = "#FF4D4E"
-    public static let colorFunctionalScarlet500: ColorRawToken = "#FF1A1B"
-    public static let colorFunctionalScarlet600: ColorRawToken = "#EA0305"
-    public static let colorFunctionalScarlet700: ColorRawToken = "#B20002"
-    public static let colorFunctionalScarlet800: ColorRawToken = "#800001"
-    public static let colorFunctionalScarlet900: ColorRawToken = "#4D0001"
-
-    // MARK: Primitive token - Colors - Functional sun
-
-    public static let colorFunctionalSun100: ColorRawToken = "#FFF7D6"
-    public static let colorFunctionalSun200: ColorRawToken = "#FFED99"
-    public static let colorFunctionalSun300: ColorRawToken = "#FFE270"
-    public static let colorFunctionalSun400: ColorRawToken = "#FFD73D"
-    public static let colorFunctionalSun500: ColorRawToken = "#FFD0D0"
-    public static let colorFunctionalSun600: ColorRawToken = "#D6AA00"
-    public static let colorFunctionalSun700: ColorRawToken = "#A38200"
-    public static let colorFunctionalSun800: ColorRawToken = "#665100"
-    public static let colorFunctionalSun900: ColorRawToken = "#3D3100"
-
-    // MARK: Primitive token - Colors - Functional malachite
-
-    public static let colorFunctionalMalachite100: ColorRawToken = "#EDFCF0"
-    public static let colorFunctionalMalachite200: ColorRawToken = "#C1F6CA"
-    public static let colorFunctionalMalachite300: ColorRawToken = "#94F0A4"
-    public static let colorFunctionalMalachite400: ColorRawToken = "#67E97E"
-    public static let colorFunctionalMalachite500: ColorRawToken = "#3DE35A"
-    public static let colorFunctionalMalachite600: ColorRawToken = "#1ECD3C"
-    public static let colorFunctionalMalachite700: ColorRawToken = "#17A02F"
-    public static let colorFunctionalMalachite800: ColorRawToken = "#0E621D"
-    public static let colorFunctionalMalachite900: ColorRawToken = "#0A4715"
-
-    // MARK: Primitive token - Colors - Functional dodger blue
-
     public static let colorFunctionalDodgerBlue100: ColorRawToken = "#F0FAFF"
     public static let colorFunctionalDodgerBlue200: ColorRawToken = "#BDE7FF"
     public static let colorFunctionalDodgerBlue300: ColorRawToken = "#8AD5FF"
@@ -101,81 +90,46 @@ extension ColorRawTokens {
     public static let colorFunctionalDodgerBlue700: ColorRawToken = "#007ABD"
     public static let colorFunctionalDodgerBlue800: ColorRawToken = "#00598A"
     public static let colorFunctionalDodgerBlue900: ColorRawToken = "#003857"
-
-    // MARK: Primitive token - Colors - Orange - Emerald
-
-    public static let colorDecorativeEmerald100: ColorRawToken = "#E5F5ED"
-    public static let colorDecorativeEmerald200: ColorRawToken = "#C0E8DA"
-    public static let colorDecorativeEmerald300: ColorRawToken = "#9BDABA"
-    public static let colorDecorativeEmerald400: ColorRawToken = "#75CCA1"
-    public static let colorDecorativeEmerald500: ColorRawToken = "#50BE87"
-    public static let colorDecorativeEmerald600: ColorRawToken = "#3BA06E"
-    public static let colorDecorativeEmerald700: ColorRawToken = "#2E7B54"
-    public static let colorDecorativeEmerald800: ColorRawToken = "#20563B"
-    public static let colorDecorativeEmerald900: ColorRawToken = "#123021"
-
-    // MARK: Primitive token - Colors - Orange - Sky
-
-    public static let colorDecorativeSky100: ColorRawToken = "#D2ECF9"
-    public static let colorDecorativeSky200: ColorRawToken = "#A5DAF3"
-    public static let colorDecorativeSky300: ColorRawToken = "#79C7EC"
-    public static let colorDecorativeSky400: ColorRawToken = "#4AB4E6"
-    public static let colorDecorativeSky500: ColorRawToken = "#1FA2E0"
-    public static let colorDecorativeSky600: ColorRawToken = "#1982B3"
-    public static let colorDecorativeSky700: ColorRawToken = "#136186"
-    public static let colorDecorativeSky800: ColorRawToken = "#0C415A"
-    public static let colorDecorativeSky900: ColorRawToken = "#06202D"
-
-    // MARK: Primitive token - Colors - Orange - Amber
-
-    public static let colorDecorativeAmber100: ColorRawToken = "#FFF0CC"
-    public static let colorDecorativeAmber200: ColorRawToken = "#FFE199"
-    public static let colorDecorativeAmber300: ColorRawToken = "#FFD266"
-    public static let colorDecorativeAmber400: ColorRawToken = "#FFC333"
-    public static let colorDecorativeAmber500: ColorRawToken = "#FFB400"
-    public static let colorDecorativeAmber600: ColorRawToken = "#CC9000"
-    public static let colorDecorativeAmber700: ColorRawToken = "#996C00"
-    public static let colorDecorativeAmber800: ColorRawToken = "#664800"
-    public static let colorDecorativeAmber900: ColorRawToken = "#332400"
-
-    // MARK: Primitive token - Colors - Orange - Amethyst
-
-    public static let colorDecorativeAmethyst100: ColorRawToken = "#F1ECF9"
-    public static let colorDecorativeAmethyst200: ColorRawToken = "#E0D4F2"
-    public static let colorDecorativeAmethyst300: ColorRawToken = "#C5ADE6"
-    public static let colorDecorativeAmethyst400: ColorRawToken = "#A885D8"
-    public static let colorDecorativeAmethyst500: ColorRawToken = "#8D60CD"
-    public static let colorDecorativeAmethyst600: ColorRawToken = "#5B2F98"
-    public static let colorDecorativeAmethyst700: ColorRawToken = "#432371"
-    public static let colorDecorativeAmethyst800: ColorRawToken = "#2C174A"
-    public static let colorDecorativeAmethyst900: ColorRawToken = "#150B23"
-
-    // MARK: Primitive token - Colors - Orange - Shocking Pink
-
-    public static let colorDecorativeShockingPink100: ColorRawToken = "#FFE5F6"
-    public static let colorDecorativeShockingPink200: ColorRawToken = "#FFB4E6"
-    public static let colorDecorativeShockingPink300: ColorRawToken = "#FF80D4"
-    public static let colorDecorativeShockingPink400: ColorRawToken = "#FF4DC3"
-    public static let colorDecorativeShockingPink500: ColorRawToken = "#FF1AB2"
-    public static let colorDecorativeShockingPink600: ColorRawToken = "#E50099"
-    public static let colorDecorativeShockingPink700: ColorRawToken = "#B20077"
-    public static let colorDecorativeShockingPink800: ColorRawToken = "#800055"
-    public static let colorDecorativeShockingPink900: ColorRawToken = "#4D0033"
-
-    // MARK: Primitive token - Colors - Orange - Deep Peach
-
-    public static let colorDecorativeDeepPeach100: ColorRawToken = "#FBEBDF"
-    public static let colorDecorativeDeepPeach200: ColorRawToken = "#F4CFB2"
-    public static let colorDecorativeDeepPeach300: ColorRawToken = "#E3B591"
-    public static let colorDecorativeDeepPeach400: ColorRawToken = "#C19372"
-    public static let colorDecorativeDeepPeach500: ColorRawToken = "#CF7E3F"
-    public static let colorDecorativeDeepPeach600: ColorRawToken = "#AA6631"
-    public static let colorDecorativeDeepPeach700: ColorRawToken = "#7E4F2A"
-    public static let colorDecorativeDeepPeach800: ColorRawToken = "#553720"
-    public static let colorDecorativeDeepPeach900: ColorRawToken = "#2E2014"
-
-    // MARK: Primitive token - Colors - Transparent black
-
+    public static let colorFunctionalLightGray80: ColorRawToken = "#F4F4F4"
+    public static let colorFunctionalLightGray160: ColorRawToken = "#EEEEEE"
+    public static let colorFunctionalLightGray240: ColorRawToken = "#E0E0E0"
+    public static let colorFunctionalLightGray320: ColorRawToken = "#D6D6D6"
+    public static let colorFunctionalLightGray400: ColorRawToken = "#CCCCCC"
+    public static let colorFunctionalLightGray480: ColorRawToken = "#C2C2C2"
+    public static let colorFunctionalLightGray560: ColorRawToken = "#BBBBBB"
+    public static let colorFunctionalLightGray640: ColorRawToken = "#ADADAD"
+    public static let colorFunctionalLightGray720: ColorRawToken = "#A3A3A3"
+    public static let colorFunctionalLightGray800: ColorRawToken = "#999999"
+    public static let colorFunctionalLightGray880: ColorRawToken = "#8F8F8F"
+    public static let colorFunctionalLightGray960: ColorRawToken = "#858585"
+    public static let colorFunctionalMalachite100: ColorRawToken = "#EDFCF0"
+    public static let colorFunctionalMalachite200: ColorRawToken = "#C1F6CA"
+    public static let colorFunctionalMalachite300: ColorRawToken = "#94F0A4"
+    public static let colorFunctionalMalachite400: ColorRawToken = "#67E97E"
+    public static let colorFunctionalMalachite500: ColorRawToken = "#3DE35A"
+    public static let colorFunctionalMalachite600: ColorRawToken = "#1ECD3C"
+    public static let colorFunctionalMalachite700: ColorRawToken = "#17A02F"
+    public static let colorFunctionalMalachite800: ColorRawToken = "#0E621D"
+    public static let colorFunctionalMalachite900: ColorRawToken = "#0A4715"
+    public static let colorFunctionalScarlet100: ColorRawToken = "#FFE5E6"
+    public static let colorFunctionalScarlet200: ColorRawToken = "#FFB2B3"
+    public static let colorFunctionalScarlet300: ColorRawToken = "#FF8081"
+    public static let colorFunctionalScarlet400: ColorRawToken = "#FF4D4E"
+    public static let colorFunctionalScarlet500: ColorRawToken = "#FF1A1B"
+    public static let colorFunctionalScarlet600: ColorRawToken = "#EA0305"
+    public static let colorFunctionalScarlet700: ColorRawToken = "#B20002"
+    public static let colorFunctionalScarlet800: ColorRawToken = "#800001"
+    public static let colorFunctionalScarlet900: ColorRawToken = "#4D0001"
+    public static let colorFunctionalSun100: ColorRawToken = "#FFF7D6"
+    public static let colorFunctionalSun200: ColorRawToken = "#FFED99"
+    public static let colorFunctionalSun300: ColorRawToken = "#FFE270"
+    public static let colorFunctionalSun400: ColorRawToken = "#FFD73D"
+    public static let colorFunctionalSun500: ColorRawToken = "#FFD000"
+    public static let colorFunctionalSun600: ColorRawToken = "#D6AA00"
+    public static let colorFunctionalSun700: ColorRawToken = "#A38200"
+    public static let colorFunctionalSun800: ColorRawToken = "#665100"
+    public static let colorFunctionalSun900: ColorRawToken = "#3D3100"
+    public static let colorFunctionalWhite: ColorRawToken = "#FFFFFF"
     public static let colorTransparentBlack0: ColorRawToken = apply(opacity: OpacityRawTokens.opacity0, on: colorFunctionalBlack)
     public static let colorTransparentBlack100: ColorRawToken = apply(opacity: OpacityRawTokens.opacity100, on: colorFunctionalBlack)
     public static let colorTransparentBlack200: ColorRawToken = apply(opacity: OpacityRawTokens.opacity200, on: colorFunctionalBlack)
@@ -186,9 +140,6 @@ extension ColorRawTokens {
     public static let colorTransparentBlack700: ColorRawToken = apply(opacity: OpacityRawTokens.opacity700, on: colorFunctionalBlack)
     public static let colorTransparentBlack800: ColorRawToken = apply(opacity: OpacityRawTokens.opacity800, on: colorFunctionalBlack)
     public static let colorTransparentBlack900: ColorRawToken = apply(opacity: OpacityRawTokens.opacity900, on: colorFunctionalBlack)
-
-    // MARK: Primitive token - Colors - Transparent white
-
     public static let colorTransparentWhite0: ColorRawToken = apply(opacity: OpacityRawTokens.opacity0, on: colorFunctionalWhite)
     public static let colorTransparentWhite100: ColorRawToken = apply(opacity: OpacityRawTokens.opacity100, on: colorFunctionalWhite)
     public static let colorTransparentWhite200: ColorRawToken = apply(opacity: OpacityRawTokens.opacity200, on: colorFunctionalWhite)
