@@ -11,30 +11,9 @@
 // Software description: A SwiftUI components library with code examples for Orange Unified Design System
 //
 
-// ଘ( ･ω･)_/ﾟ･:*:･｡☆
-// [File to generate with the tokenator]
-
 // swiftlint:disable missing_docs
 
-/// Extracted in a separated file to help the *Figma* JSON to Swift parser to generate files to include easily.
 extension ElevationRawTokens {
-
-    // MARK: Primitive token - Elevation - X
-
-    public static let elevationX0: ElevationRawToken = 0
-
-    // MARK: Primitive token - Elevation - Y
-
-    public static let elevationY0: ElevationRawToken = 0
-    public static let elevationY100: ElevationRawToken = 1
-    public static let elevationY200: ElevationRawToken = 2
-    public static let elevationY300: ElevationRawToken = 4
-    public static let elevationY400: ElevationRawToken = 8
-    public static let elevationY500: ElevationRawToken = 12
-    public static let elevationY600: ElevationRawToken = 20
-
-    // MARK: Primitive token - Elevation - Blur
-
     public static let elevationBlur0: ElevationRawToken = 0
     public static let elevationBlur100: ElevationRawToken = 1
     public static let elevationBlur200: ElevationRawToken = 2
@@ -43,6 +22,14 @@ extension ElevationRawTokens {
     public static let elevationBlur500: ElevationRawToken = 8
     public static let elevationBlur600: ElevationRawToken = 12
     public static let elevationBlur700: ElevationRawToken = 20
+    public static let elevationX0: ElevationRawToken = 0
+    public static let elevationY0: ElevationRawToken = 0
+    public static let elevationY100: ElevationRawToken = 1
+    public static let elevationY200: ElevationRawToken = 2
+    public static let elevationY300: ElevationRawToken = 4
+    public static let elevationY400: ElevationRawToken = 8
+    public static let elevationY500: ElevationRawToken = 12
+    public static let elevationY600: ElevationRawToken = 20
 }
 
 // swiftlint:enable missing_docs
