@@ -12,11 +12,13 @@
 //
 
 // ଘ( ･ω･)_/ﾟ･:*:･｡☆
-// [File to generate with the tokenator]
+// [File to generate with the tokenator with Figma able to output composites and tokenatoer able to manage them]
+// WARNING: Not synchronized anymore with the Figjam / Figma by developers team
+// Create an issue for update https://github.com/Orange-OpenSource/ouds-ios/issues/new?template=token_update.yml
 
 // swiftlint:disable missing_docs
 
-/// Extracted in a separated file because the the *Figma* JSON to Swift parser is not abale to process such composite tokens, and the *Figma* tool is not able to.
+/// Extracted in a separated file because the  *Figma* JSON to Swift, i.e. the *tokenator* parser is not able to process such composite tokens, and the *Figma* tool is not able to.
 ///
 /// **Beware, may result in desynchronization between generated raw tokens values and theses composite raw tokens**
 extension TypographyRawTokens {
