@@ -13,6 +13,8 @@
 
 // ଘ( ･ω･)_/ﾟ･:*:･｡☆
 // [File to generate with the tokenator]
+// WARNING: Not synchronized anymore with the Figjam / Figma by developers team
+// Create an issue for update https://github.com/Orange-OpenSource/ouds-ios/issues/new?template=token_update.yml
 
 // swiftlint:disable missing_docs
 
@@ -30,7 +32,8 @@ public protocol BorderSemanticTokens {
     var borderWidthMedium: BorderWidthSemanticToken { get }
     var borderWidthThick: BorderWidthSemanticToken { get }
     var borderWidthThicker: BorderWidthSemanticToken { get }
-    var borderWidthOutsideFocus: BorderWidthSemanticToken { get }
+    var borderWidthFocus: BorderWidthSemanticToken { get }
+    var borderWidthFocusInset: BorderWidthSemanticToken { get }
 
     // MARK: - Semantic token - Border - Radius
 
