@@ -13,12 +13,17 @@
 
 // ଘ( ･ω･)_/ﾟ･:*:･｡☆
 // [File to generate with the tokenator]
+// WARNING: Not synchronized anymore with the Figjam / Figma by developers team
+// Create an issue for update https://github.com/Orange-OpenSource/ouds-ios/issues/new?template=token_update.yml
 
 // swiftlint:disable missing_docs
 
 /// This is a group of semantic tokens for **elevations**.
-/// It defines all elevation semantic tokens a theme must have. (`ElevationXSemanticToken`,`ElevationYSemanticToken`,
-/// `ElevationBlurSemanticToken` and `ElevationColorSemanticToken`)
+/// It defines all elevation semantic tokens a theme must have (`ElevationXSemanticToken`,`ElevationYSemanticToken`,
+/// `ElevationBlurSemanticToken` and `ElevationColorSemanticToken`).
+///
+/// However the composite tokens (here for bow shadows) are defined in `ElevationCompositeSemanticTokens` because the *tokenator*
+/// is not able to generate them yet, and they must be defined elsewhere to not be deleted.
 ///
 /// In the future this file should be generated by a parser converting Figma JSON to Swift code.
 public protocol ElevationSemanticTokens {
@@ -33,7 +38,6 @@ public protocol ElevationSemanticTokens {
     var elevationXStickyDefault: ElevationXSemanticToken { get }
     var elevationXStickyEmphasized: ElevationXSemanticToken { get }
     var elevationXStickyNavigationScrolled: ElevationXSemanticToken { get }
-    var elevationXFocus: ElevationXSemanticToken { get }
 
     // MARK: Semantic token - Elevation - Y
 
@@ -45,7 +49,6 @@ public protocol ElevationSemanticTokens {
     var elevationYStickyDefault: ElevationYSemanticToken { get }
     var elevationYStickyEmphasized: ElevationYSemanticToken { get }
     var elevationYStickyNavigationScrolled: ElevationYSemanticToken { get }
-    var elevationYFocus: ElevationYSemanticToken { get }
 
     // MARK: Semantic token - Elevation - Blur
 
@@ -57,7 +60,6 @@ public protocol ElevationSemanticTokens {
     var elevationBlurStickyDefault: ElevationBlurSemanticToken { get }
     var elevationBlurStickyEmphasized: ElevationBlurSemanticToken { get }
     var elevationBlurStickyNavigationScrolled: ElevationBlurSemanticToken { get }
-    var elevationBlurFocus: ElevationBlurSemanticToken { get }
 
     // MARK: Semantic token - Elevation - Color
 
@@ -69,19 +71,6 @@ public protocol ElevationSemanticTokens {
     var elevationColorStickyDefault: ElevationColorSemanticToken { get }
     var elevationColorStickyEmphasized: ElevationColorSemanticToken { get }
     var elevationColorStickyNavigationScrolled: ElevationColorSemanticToken { get }
-    var elevationColorFocus: ElevationColorSemanticToken { get }
-
-    // MARK: Semantic token - Elevation - Box shadow
-
-    var elevationNone: ElevationCompositeSemanticToken { get }
-    var elevationRaised: ElevationCompositeSemanticToken { get }
-    var elevationDrag: ElevationCompositeSemanticToken { get }
-    var elevationOverlayDefault: ElevationCompositeSemanticToken { get }
-    var elevationOverlayEmphasized: ElevationCompositeSemanticToken { get }
-    var elevationStickyDefault: ElevationCompositeSemanticToken { get }
-    var elevationStickyEmphasized: ElevationCompositeSemanticToken { get }
-    var elevationStickyNavigationScrolled: ElevationCompositeSemanticToken { get }
-    var elevationFocus: ElevationCompositeSemanticToken { get }
 }
 
 // swiftlint:enable missing_docs
