@@ -11,10 +11,10 @@
 // Software description: A SwiftUI components library with code examples for Orange Unified Design System 
 //
 
-import OUDSTokensRaw
-
 // ଘ( ･ω･)_/ﾟ･:*:･｡☆
 // [File to generate with the tokenator]
+// WARNING: Not synchronized anymore with the Figjam / Figma by developers team
+// Create an issue for update https://github.com/Orange-OpenSource/ouds-ios/issues/new?template=token_update.yml
 
 // swiftlint:disable missing_docs
 
@@ -22,12 +22,15 @@ import OUDSTokensRaw
 /// It defines all typography semantic tokens a theme must have
 /// (`TypographyFontFamilySemanticToken`, `TypographyFontWeightSemanticToken`, `TypographyFontSizeSemanticToken`, `TypographyFontLineHeightSemanticToken`).
 ///
+/// However the composite tokens (here the ones gathering each type of semantic token here) are defined in `TypographyCompositeSemanticTokens` because the *tokenator*
+/// is not able to generate them yet, and they must be defined elsewhere to not be deleted.
+///
 /// In the future this file should be generated by a parser converting Figma JSON to Swift code.
 public protocol TypographySemanticTokens {
 
     // MARK: - Semantic token - Typography - Font - Family
 
-    var fontFamily: TypographyFontFamilyRawToken { get }
+    var fontFamily: TypographyFontFamilySemanticToken { get }
     var fontFamilyDisplay: TypographyFontFamilySemanticToken { get }
     var fontFamilyHeading: TypographyFontFamilySemanticToken { get }
     var fontFamilyBody: TypographyFontFamilySemanticToken { get }
@@ -37,6 +40,7 @@ public protocol TypographySemanticTokens {
     // MARK: - Semantic token - Typography - Font - Weight
 
     var fontWeightDefault: TypographyFontWeightSemanticToken { get }
+    var fontWeightCode: TypographyFontWeightSemanticToken { get }
     var fontWeightStrong: TypographyFontWeightSemanticToken { get }
     var fontWeightDisplay: TypographyFontWeightSemanticToken { get }
     var fontWeightHeading: TypographyFontWeightSemanticToken { get }
@@ -44,7 +48,6 @@ public protocol TypographySemanticTokens {
     var fontWeightBodyStrong: TypographyFontWeightSemanticToken { get }
     var fontWeightLabelDefault: TypographyFontWeightSemanticToken { get }
     var fontWeightLabelStrong: TypographyFontWeightSemanticToken { get }
-    var fontWeightCode: TypographyFontWeightSemanticToken { get }
 
     // MARK: - Semantic token - Typography - Font - Size
 
@@ -106,44 +109,6 @@ public protocol TypographySemanticTokens {
     var fontLetterSpacingLabelSmall: MultipleFontLetterSpacingTokens { get }
     var fontLetterSpacingCodeMedium: MultipleFontLetterSpacingTokens { get }
     var fontLetterSpacingCodeSmall: MultipleFontLetterSpacingTokens { get }
-
-    // MARK: - Semantic tokens - Typography - Composites - Display
-
-    var typeDisplayLarge: MultipleTypographyTokens { get }
-    var typeDisplayMedium: MultipleTypographyTokens { get }
-    var typeDisplaySmall: MultipleTypographyTokens { get }
-
-    // MARK: - Semantic tokens - Typography - Composites - Heading
-
-    var typeHeadingXLarge: MultipleTypographyTokens { get }
-    var typeHeadingLarge: MultipleTypographyTokens { get }
-    var typeHeadingMedium: MultipleTypographyTokens { get }
-    var typeHeadingSmall: MultipleTypographyTokens { get }
-
-    // MARK: - Semantic tokens - Typography - Composites - Body
-
-    var typeBodyDefaultLarge: MultipleTypographyTokens { get }
-    var typeBodyDefaultMedium: MultipleTypographyTokens { get }
-    var typeBodyDefaultSmall: MultipleTypographyTokens { get }
-    var typeBodyStrongLarge: MultipleTypographyTokens { get }
-    var typeBodyStrongMedium: MultipleTypographyTokens { get }
-    var typeBodyStrongSmall: MultipleTypographyTokens { get }
-
-    // MARK: - Semantic tokens - Typography - Composites - Label
-
-    var typeLabelDefaultXLarge: MultipleTypographyTokens { get }
-    var typeLabelDefaultLarge: MultipleTypographyTokens { get }
-    var typeLabelDefaultMedium: MultipleTypographyTokens { get }
-    var typeLabelDefaultSmall: MultipleTypographyTokens { get }
-    var typeLabelStrongXLarge: MultipleTypographyTokens { get }
-    var typeLabelStrongLarge: MultipleTypographyTokens { get }
-    var typeLabelStrongMedium: MultipleTypographyTokens { get }
-    var typeLabelStrongSmall: MultipleTypographyTokens { get }
-
-    // MARK: - Semantic tokens - Typography - Composites - Code
-
-    var typeCodeMedium: MultipleTypographyTokens { get }
-    var typeCodeSmall: MultipleTypographyTokens { get }
 }
 
 // swiftlint:enable missing_docs
