@@ -47,7 +47,7 @@ final class MultipleTypographyTokensTests: XCTestCase {
         let third = MultipleTypographyTokens(compact: TypographyRawTokens.typeRegular150, regular: TypographyRawTokens.typeBold750)
         let fourth = MultipleTypographyTokens(compact: TypographyRawTokens.typeBold850, regular: TypographyRawTokens.typeBold550)
         let fifth = MultipleTypographyTokens(compact: TypographyRawTokens.typeRegular150, regular: TypographyRawTokens.typeBold550)
-        let sixth = MultipleSizingTokens(compact: 0, regular: 0)
+        let sixth = MultipleSizeTokens(compact: 0, regular: 0)
 
         XCTAssertTrue(first.isEqual(first))
         XCTAssertFalse(first.isEqual(second))
