@@ -36,14 +36,14 @@ extension XCTestCase {
     public func XCTAssertColorLighterThan(_ left: String, _ right: String, file: StaticString = #file, line: UInt = #line) {
         let leftBrightness = hexadecimalStringToInt(left)
         let rightBrightness = hexadecimalStringToInt(right)
-        XCTAssertGreaterThan(leftBrightness, rightBrightness, "Right color is not lighter than the left color", file: file, line: line)
+        XCTAssertGreaterThan(leftBrightness, rightBrightness, "Right color is not lighter than the left color", file: (file), line: line)
     }
 
     /// Compares two colors in hexadecimal values and asserts if `left` is "darker" than `right`, i.e. has a lower integer value
     public func XCTAssertColorDarkerThan(_ left: String, _ right: String, file: StaticString = #file, line: UInt = #line) {
         let leftBrightness = hexadecimalStringToInt(left)
         let rightBrightness = hexadecimalStringToInt(right)
-        XCTAssertLessThan(leftBrightness, rightBrightness, "Right color is not darker than the left color", file: file, line: line)
+        XCTAssertLessThan(leftBrightness, rightBrightness, "Right color is not darker than the left color", file: (file), line: line)
     }
 
     // MARK: - Helpers
