@@ -45,7 +45,7 @@ struct ElevationTokenPage: View {
         return ShowcaseTokenIllustration(tokenName: name, tokenValue: value) {
             Rectangle()
                 .frame(width: theme.sizeIconDecorativeTallest, height: theme.sizeIconDecorativeTallest)
-                .foregroundColor(theme.colorBackgroundSecondary.color(for: colorScheme))
+                .foregroundColor(theme.colorBgSecondary.color(for: colorScheme))
                 .shadow(elevation: token)
                 .padding(.bottom, 2)
         }
@@ -56,7 +56,6 @@ struct ElevationTokenPage: View {
 
 private enum NamedElevation: String, CaseIterable {
     case elevationNone
-    case elevationFocus
     case elevationRaised
     case elevationStickyNavigationScrolled
     case elevationOverlayDefault
@@ -69,8 +68,6 @@ private enum NamedElevation: String, CaseIterable {
         switch self {
         case .elevationNone:
             return theme.elevationNone
-        case .elevationFocus:
-            return theme.elevationFocus
         case .elevationRaised:
             return theme.elevationRaised
         case .elevationStickyNavigationScrolled:
